hosted, db changed to the hosted one, Price changed to decimal

+++ MyFirstASP.NET/Data/Models/Car.cs
@@ -12,7 +12,7 @@ namespace MyFirstASP.NET.Data.Models
         public string ShortDesc { get; set; }
         public string LongDesc { get; set; }
         public string Img { get; set; }
-        public ushort Price { get; set; }
+        public decimal Price { get; set; }
         public bool IsFavourite { get; set; }
         public bool Available { get; set; }
         public int CategoryId { get; set; }
